added getNotesRouter, moved security file.
