+++ DOCUMENTATION.md
@@ -4,14 +4,15 @@ This is a RESTful API to connect the Lambda Notes Project by AJ Genung to the Da
 
 ## API Endpoints
 
-| Type   | Path             | Data                   |
-| ------ | ---------------- | ---------------------- |
-| GET    | /api/logout      | N/A                    |
-| POST   | /api/login       | username, password     |
-| POST   | /api/register    | username, password     |
-| POST   | /api/newnote     | title, body            |
-| PUT    | /api/updatenote  | note.\_id, title, body |
-| DELETE | /api/destroynote | note.\_id              |
+| Type   | Path             | Data                         |
+| ------ | ---------------- | ---------------------------- |
+| GET    | /api/logout      | N/A                          |
+| POST   | /api/login       | username, password           |
+| POST   | /api/register    | username, password           |
+| GET    | /api/getnotes    | JWT on headers.Authorization |
+| POST   | /api/newnote     | title, body                  |
+| PUT    | /api/updatenote  | note.\_id, title, body       |
+| DELETE | /api/destroynote | note.\_id                    |
 
 ## Route Details
 
@@ -31,15 +32,19 @@ This is a RESTful API to connect the Lambda Notes Project by AJ Genung to the Da
 
 ### Note Manipulation
 
-1.  [POST] request to `/api/newnote` requires a title and active `JWT`. `body` is optional and will default to 'Default Entry' if not provided. `{ JWT, title, body }`
+1.  [GET] request to `/api/getnotes` requires active JWT from current active user to be present on `headers.Authorization`.
+
+    * Will return user object with `{ _id, username, notes }`
+
+2.  [POST] request to `/api/newnote` requires a title and active `JWT`. `body` is optional and will default to 'Default Entry' if not provided. `{ JWT, title, body }`
 
     * Saves note to the database, and returns an array of all notes from active user.
 
-2.  [PUT] request to `/api/updatenote` requires active `JWT` and `note._id`. `title` and `body` are optional, but if your updating something, you should probably send at least one of these...
+3.  [PUT] request to `/api/updatenote` requires active `JWT` and `note._id`. `title` and `body` are optional, but if your updating something, you should probably send at least one of these...
 
     * Updates the note on the database and returns an array of all notes from the active user.
 
-3.  [DELETE] request to `/api/destroynote/:id` requires active `JWT` on the body, and the `:id` param must be an active note.\_id.
+4.  [DELETE] request to `/api/destroynote/:id` requires active `JWT` on the body, and the `:id` param must be an active note.\_id.
 
     * Deletes the note from the database, and returns an array of all the notes from the active user.
