+++ Note/getNotesRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Note = require('./Note');
-const { restricted } = require('../User/security');
+const { restricted } = require('../secrets/security');
 
 router.get('/', restricted, (req, res) => {
   res.json({ user: req.user });
