+++ User/loginRouter.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const User = require('./User');
 // const { makeToken } = require('./auth');
-const { authenticate, makeToken } = require('./security');
+const { authenticate, makeToken } = require('../secrets/security');
 
 router.post('/', authenticate, (req, res) => {
   // const { username, password } = req.body;
